step and check display
